add minecraft datatype file for minecraft specific packets
